requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
mysql-connector-python==8.1.0
pymysql>=1.1.0
cryptography>=3.0.0
